Refactor in helper package
