+++ onion.go
@@ -8,6 +8,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/goraz/onion/helper"
 	"github.com/mitchellh/mapstructure"
 )
 
@@ -484,37 +485,9 @@ func (o *Onion) LayersData() []map[string]interface{} {
 // MergedLayersData is used to get all layers data merged into one
 // Latest added overwrite previous ones.
 func (o *Onion) MergedLayersData() map[string]interface{} {
-	layers := o.LayersData()
+	layersData := o.LayersData()
 
-	mergedLayer := layers[len(layers)-1]
-	layers = layers[:len(layers)-1]
-
-	for i := len(layers) - 1; i >= 0; i-- {
-		mergedLayer = mergeKeys(mergedLayer, layers[i])
-	}
-
-	return mergedLayer
-}
-
-// mergeKeys recursively merge right into left, never replacing any key that already exists in left
-func mergeKeys(left, right map[string]interface{}) map[string]interface{} {
-	if left == nil {
-		return right
-	}
-
-	for key, rightVal := range right {
-		if leftVal, present := left[key]; present {
-			_, leftValIsAMap := leftVal.(map[string]interface{})
-			_, rightValIsAMap := leftVal.(map[string]interface{})
-
-			if leftValIsAMap && rightValIsAMap {
-				left[key] = mergeKeys(leftVal.(map[string]interface{}), rightVal.(map[string]interface{}))
-			}
-		} else {
-			left[key] = rightVal
-		}
-	}
-	return left
+	return helper.MergeLayersData(layersData)
 }
 
 // MergeAndDecode try to convert merged layers in the output structure.
